pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tradingview-scraper"
version = "1.0.0"
description = "A powerful Python library for scraping real-time market data, indicators, and ideas from TradingView."
authors = [
    {name = "Smit Kunpara", email = "smitkunpara@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["tradingview", "scraper", "python", "finance", "market-data", "technical-analysis", "real-time"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "requests>=2.32.4",
    "websocket-client>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock",
    "pytest-cov",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]
csv = [
    "pandas>=2.0.3",
]

[project.urls]
Homepage = "https://github.com/smitkunpara/tv-scraper"
Repository = "https://github.com/smitkunpara/tv-scraper"
Issues = "https://github.com/smitkunpara/tv-scraper/issues"

[tool.hatch.build.targets.wheel]
packages = ["tv_scraper"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "live: marks tests that require live network connections to TradingView (deselect with '-m \"not live\"')",
]
